feat(Authentication): enhance authentication services in Program.cs
This commit enhances the authentication services in the Program.cs file. The added code includes the use of Blazored LocalStorage, the addition of IAuthService and AuthenticationStateProvider to the service collection, and the configuration of JwtBearer authentication. Also, the global using directives for PacificToursApp.Client.Services.AuthService and Microsoft.AspNetCore.Components.Authorization have been added.